+++ examples/app/src/main/java/de/infolektuell/bass/app/VersionNumber.java
@@ -1,8 +1,0 @@
-package de.infolektuell.bass.app;
-
-public record VersionNumber(int major, int minor, int patch, int build) {
-    @Override
-    public String toString() {
-        return major + "." + minor + "." + patch + "." + build;
-    }
-}