+++ examples/nativelib/src/main/cpp/hello.cpp
@@ -1,0 +1,8 @@
+#include "hello.h"
+
+extern "C" {
+    const char *hello_world(void)
+    {
+        return "Hello World";
+    }
+}